+++ packages/terra-date-picker/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Added the break line tag in datePickerField between help message.
+
 ## 4.97.0 - (January 25, 2024)
 
 * Changed